+++ NEWS.md
@@ -1,3 +1,5 @@
+# rlang (development version)
+
 # rlang 1.0.0
 
 ## Major changes